Feature/cart token userid

+++ src/main/java/com/invoice/api/service/SvcCartItemImp.java
@@ -8,10 +8,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.invoice.api.dto.In.DtoCartItemIn;
 import com.invoice.api.dto.In.DtoCartItemQuantityIn;
 import com.invoice.api.entity.CartItem;
 import com.invoice.api.repository.RepoCartItem;
 import com.invoice.common.dto.ApiResponse;
+import com.invoice.common.mapper.MapperCartItem;
+import com.invoice.common.util.JwtDecoder;
 import com.invoice.exception.ApiException;
 import com.invoice.exception.DBAccessException;
 
@@ -21,24 +24,37 @@ public class SvcCartItemImp implements SvcCartItem {
     @Autowired
     private RepoCartItem repo;
 
+    @Autowired
+    private MapperCartItem mapper;
+
+    @Autowired
+	private JwtDecoder jwtDecoder;
+
     @Override
-    public ResponseEntity<List<CartItem>> getCartItems(Integer customerId) {
+    public ResponseEntity<List<CartItem>> getCartItems() {
         try {
-            List<CartItem> cartItems = repo.findByCustomerId(customerId);
-            return new ResponseEntity<>(cartItems, HttpStatus.OK);
+            if(jwtDecoder.isAdmin()) {
+				return new ResponseEntity<>(repo.findAll(), HttpStatus.OK);
+			}else {
+				Integer userId = jwtDecoder.getUserId();
+				return new ResponseEntity<>(repo.findByUserId(userId), HttpStatus.OK);
+			}
         } catch (DataAccessException e) {
             throw new DBAccessException(e);
         }
     }
 
     @Override
-    public ResponseEntity<ApiResponse> createCartItem(CartItem in) {
+    public ResponseEntity<ApiResponse> createCartItem(DtoCartItemIn in) {
         try {
-            if (in.getCartItemId() != null)
-                throw new ApiException(HttpStatus.BAD_REQUEST, "El id es generado automáticamente");
-            // Guardar el nuevo CartItem
+            if(!jwtDecoder.isAdmin() && jwtDecoder.getUserId() != in.getUserId()) {
+				throw new ApiException(HttpStatus.FORBIDDEN, "El token no es valido para agregar articulo a este carrito");
+			} else if (jwtDecoder.isAdmin() && jwtDecoder.getUserId() == in.getUserId()) {
+                throw new ApiException(HttpStatus.FORBIDDEN, "El token de administrador no esta asociado a ningún carrito");
+            }
 
-            repo.save(in);
+            CartItem cartItem = mapper.fomCartItem(in);
+            repo.save(cartItem);
             // Retornar respuesta exitosa con código 201 (CREATED)
             return new ResponseEntity<>(new ApiResponse("El artículo ha sido agregado al carrito exitosamente"),
                     HttpStatus.CREATED);
@@ -58,6 +74,10 @@ public class SvcCartItemImp implements SvcCartItem {
         try {
             CartItem cartItem = validateCartItemId(cartItemId);
 
+            if(!jwtDecoder.isAdmin() && jwtDecoder.getUserId() != cartItem.getUserId()) {
+				throw new ApiException(HttpStatus.FORBIDDEN, "El token no es valido para actualizar este articulo");
+			}
+
             cartItem.setQuantity(cartItem.getQuantity() + in.getQuantity());
 
             if (cartItem.getQuantity() <= 0) {
@@ -76,6 +96,12 @@ public class SvcCartItemImp implements SvcCartItem {
     @Override
     public ResponseEntity<ApiResponse> deleteCartItem(String cartItemId) {
         try {
+            CartItem cartItem = validateCartItemId(cartItemId);
+
+            if(!jwtDecoder.isAdmin() && jwtDecoder.getUserId() != cartItem.getUserId()) {
+				throw new ApiException(HttpStatus.FORBIDDEN, "El token no es valido para eliminar este articulo");
+			}
+
             repo.deleteById(cartItemId);
 
             return new ResponseEntity<>(new ApiResponse("El articulo ha sido eliminado"), HttpStatus.OK);
@@ -85,16 +111,13 @@ public class SvcCartItemImp implements SvcCartItem {
     }
 
     @Override
-    public ResponseEntity<ApiResponse> deleteCartItems(Integer customerId) {
+    public ResponseEntity<ApiResponse> deleteCartItems() {
         try {
-            // Eliminar todos los CartItem con el customerId dado
-            repo.deleteByCustomerId(customerId);
-
-            // Devolver respuesta de éxito con código 200 (OK)
+            repo.deleteByUserId(jwtDecoder.getUserId());
+            
             return new ResponseEntity<>(new ApiResponse("Los artículos del carrito han sido eliminados"),
                     HttpStatus.OK);
         } catch (DataAccessException e) {
-            // En caso de error en acceso a datos, lanzar excepción personalizada
             throw new DBAccessException(e);
         }
     }
